Add README, tests, and timezone input
